Split race and demographics form selectors
