+++ spec/javascripts/selectors/screening/peopleFormSelectorSpec.js
@@ -2,11 +2,6 @@ import {fromJS, List, Seq} from 'immutable'
 import {
   getFilteredPersonRolesSelector,
   getPeopleWithEditsSelector,
-  getPersonDemographicsSelector,
-  getPersonRacesSelector,
-  getPersonRaceDetailsSelector,
-  getIsApproximateAgeDisabledSelector,
-  getIsRaceIndeterminateValueSelector,
   getErrorsSelector,
   getNamesRequiredSelector,
   getPersonAlertErrorMessageSelector,
@@ -449,169 +444,6 @@ describe('peopleFormSelectors', () => {
           {label: 'Anonymous Reporter', value: 'Anonymous Reporter', disabled: false},
         ]))
       })
-    })
-  })
-
-  describe('getIsApproximateAgeDisabledSelector', () => {
-    it('is set to false if the given person does not have a date of birth', () => {
-      const peopleForm = {1: {date_of_birth: {value: null}}}
-      const state = fromJS({peopleForm})
-      expect(getIsApproximateAgeDisabledSelector(state, '1')).toBe(false)
-    })
-
-    it('is set to true if the given person has a date of birth', () => {
-      const peopleForm = {1: {date_of_birth: {value: '13/0/-514'}}}
-      const state = fromJS({peopleForm})
-      expect(getIsApproximateAgeDisabledSelector(state, '1')).toBe(true)
-    })
-  })
-
-  describe('getPersonDemographicsSelector', () => {
-    it('returns a blank person when person does not exist', () => {
-      const peopleForm = {1: {}}
-      const state = fromJS({peopleForm})
-      expect(getPersonDemographicsSelector(state, '2').toJS()).toEqual({
-        approximateAge: '',
-        approximateAgeUnit: 'years',
-        dateOfBirth: '',
-        gender: '',
-        languages: [],
-      })
-    })
-
-    it('includes the approximate age for the given person', () => {
-      const peopleForm = {1: {approximate_age: {value: '1'}}}
-      const state = fromJS({peopleForm})
-      expect(getPersonDemographicsSelector(state, '1').get('approximateAge')).toEqual('1')
-    })
-
-    it('includes the approximate age unit for the given person', () => {
-      const peopleForm = {1: {approximate_age_units: {value: 'year'}}}
-      const state = fromJS({peopleForm})
-      expect(getPersonDemographicsSelector(state, '1').get('approximateAgeUnit')).toEqual('year')
-    })
-
-    it('includes the date of birth for the given person', () => {
-      const peopleForm = {1: {date_of_birth: {value: '13/0/-514'}}}
-      const state = fromJS({peopleForm})
-      expect(getPersonDemographicsSelector(state, '1').get('dateOfBirth')).toEqual('13/0/-514')
-    })
-
-    it('includes the gender for the given person', () => {
-      const peopleForm = {1: {gender: {value: 'known'}}}
-      const state = fromJS({peopleForm})
-      expect(getPersonDemographicsSelector(state, '1').get('gender')).toEqual('known')
-    })
-
-    it('includes the languages for the given person', () => {
-      const peopleForm = {1: {languages: {value: ['Ω', 'ß']}}}
-      const state = fromJS({peopleForm})
-      expect(getPersonDemographicsSelector(state, '1').get('languages'))
-        .toEqualImmutable(fromJS(['Ω', 'ß']))
-    })
-  })
-  describe('getPersonRacesSelector', () => {
-    it('returns the races for the person with the passed id', () => {
-      const peopleForm = {
-        one: {
-          races: {
-            White: {value: true},
-            Asian: {value: true},
-          },
-        },
-        two: {
-          races: {},
-        },
-      }
-      const state = fromJS({peopleForm})
-      expect(getPersonRacesSelector(state, 'one')).toEqualImmutable(fromJS({
-        White: true,
-        'Black or African American': false,
-        Asian: true,
-        'American Indian or Alaska Native': false,
-        'Native Hawaiian or Other Pacific Islander': false,
-        Unknown: false,
-        Abandoned: false,
-        'Declined to answer': false,
-      }))
-    })
-  })
-  describe('getPersonRaceDetailsSelector', () => {
-    it('returns the races for the person with the passed id', () => {
-      const peopleForm = {
-        one: {
-          race_details: {
-            White: {value: 'race_detail_1'},
-            Asian: {value: 'race_detail_2'},
-          },
-        },
-        two: {
-          race_details: {},
-        },
-      }
-      const state = fromJS({peopleForm})
-      expect(getPersonRaceDetailsSelector(state, 'one')).toEqualImmutable(fromJS({
-        White: 'race_detail_1',
-        'Black or African American': '',
-        Asian: 'race_detail_2',
-        'American Indian or Alaska Native': '',
-        'Native Hawaiian or Other Pacific Islander': '',
-        Unknown: '',
-        Abandoned: '',
-        'Declined to answer': '',
-      }))
-    })
-  })
-
-  describe('getIsRaceIndeterminateValueSelector', () => {
-    it('returns true if persons race is Unknown', () => {
-      const peopleForm = {
-        one: {
-          races: {
-            Unknown: {value: true},
-          },
-        },
-        two: {
-          races: {},
-        },
-      }
-      const state = fromJS({peopleForm})
-      expect(getIsRaceIndeterminateValueSelector(state, 'one')).toEqual(true)
-      expect(getIsRaceIndeterminateValueSelector(state, 'two')).toEqual(false)
-    })
-
-    it('returns true if persons race is Abandoned', () => {
-      const peopleForm = {
-        one: {
-          races: {
-            Abandoned: {value: true},
-          },
-        },
-        two: {
-          races: {
-            White: {value: true},
-          },
-        },
-      }
-      const state = fromJS({peopleForm})
-      expect(getIsRaceIndeterminateValueSelector(state, 'one')).toEqual(true)
-      expect(getIsRaceIndeterminateValueSelector(state, 'two')).toEqual(false)
-    })
-
-    it("returns true if persons race is 'Declined to answer'", () => {
-      const peopleForm = {
-        one: {
-          races: {
-            'Declined to answer': {value: true},
-          },
-        },
-        two: {
-          Asian: {value: true},
-        },
-      }
-      const state = fromJS({peopleForm})
-      expect(getIsRaceIndeterminateValueSelector(state, 'one')).toEqual(true)
-      expect(getIsRaceIndeterminateValueSelector(state, 'two')).toEqual(false)
     })
   })
